added random users without organisations

+++ server/src/main/java/urban/server/repositories/UserRepository.java
@@ -9,7 +9,7 @@ public interface UserRepository {
 
     void delete(User user);
 
-    User findById(int id);
+    User findById(Long id);
 
     List<User> findAll();
 }
